Consolidate serializing and deserializing CLLocation
